+++ src/qsidentify/__init__.py
@@ -1,3 +1,5 @@
 from ._version import __version__
+from .api import IdentificationResult, identify
+from .drivers import DriverInfo, drivers
 
-__all__ = ["__version__"]
+__all__ = ["DriverInfo", "IdentificationResult", "__version__", "drivers", "identify"]